registration/serializers.py
from logging import raiseExceptions
from tokenize import group
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import *
from django.contrib.auth.models import Group

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email', 'first_name', 'last_name', 'date_joined', 'last_login','password')

class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = ('__all__')

